+++ src/Topologies/interval.jl
@@ -7,7 +7,7 @@ A sequential topology on an [`Meshes.IntervalMesh`](@ref).
 """
 struct IntervalTopology{
     C <: ClimaComms.AbstractCommsContext,
-    M <: Meshes.IntervalMesh,
+    M <: Union{Meshes.IntervalMesh, Nothing},
     B,
 } <: AbstractIntervalTopology
     """
@@ -19,14 +19,6 @@ struct IntervalTopology{
 end
 
 Adapt.@adapt_structure IntervalTopology
-
-## gpu
-struct DeviceIntervalTopology{B} <: AbstractIntervalTopology
-    boundaries::B
-end
-
-ClimaComms.context(topology::DeviceIntervalTopology) = DeviceSideContext()
-ClimaComms.device(topology::DeviceIntervalTopology) = DeviceSideDevice()
 
 function IntervalTopology(
     context::ClimaComms.AbstractCommsContext,